requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
transformers==4.35.2
torch==2.9.0
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
openai
requests
python-dotenv
pgvector
langchain
langchain-community
langchain-openai
langchain-core
langchain-community
langchain-openai
langchain-core
fastapi
uvicorn[standard]
httpx
pyautogui
requests
openai
python-dotenv
sqlalchemy
psycopg2-binary